Shell completion (#62)
* comment added

* art: removed

* k8s: adding

* k8s: tests added

* k8s: tests added

* k8s: tests added

* k8s: tests added

* k8s: tests added

* k8s: delete pod added

* k8s: pod shell added

* k8s: pod shell added

* k8s: pod shell added

* k8s: pod shell added

* k8s: pod shell added

* k8s: pod shell added

* k8s: pod logs added

* k8s: pod logs added

* k8s: pod logs added

* k8s: pod shell added

* k8s: workflow added

* k8s: docs added

* k8s: docs added

* k8s: namespace added

* k8s: namespace added

* k8s: test added

* k8s: test added

* k8s: test added

* k8s: pr fixes

* k8s: pr fixes

* k8s: pr fixes

* k8s: pr fixes

* k8s: brew and docs fixes

* pr: fixes

* k8s: pod shell added

* k8s: pod shell added

* k8s: pod shell added

* Rename LICENSE.txt to LICENSE and update SECURITY.md

Renamed the license file from LICENSE.txt to LICENSE for consistency and readability. Updated SECURITY.md to reflect the support status of version 0.3.x and to discontinue support for version 0.2.x.

* Add documentation for shell completion and installation

Added detailed documentation for shell completion support in Bash and Zsh, along with installation and upgrade instructions using pip and Homebrew. Enhanced the main index page with a welcoming message, quick links, and updated the navigation structure.

* Refactor documentation structure

Deleted outdated `crypto` and `k8s` index files, and reorganized the `commands` section to include various topics like `cron`, `crypto`, `data`, `hash`, `info`, and `k8s`. This restructuring improves navigation and centralizes the command-related documentation under a single directory.

* Enhance shell completion documentation

Added examples to illustrate tab completion for commands and options in Bash and Zsh shells. Provided code snippets for better clarity on how to use the feature effectively. Reorganized existing tips for improved readability.

* Add installation section reference in documentation

Updated the `Docs` link in the main index to use the installation reference for better navigation. Also added a label for the installation section to ensure consistency and clarity.